# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayTrace Backend",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/diagnosis-runs/{run_id}": {
      "get": {
        "tags": [
          "diagnosis",
          "diagnosis"
        ],
        "summary": "Get Diagnosis Run",
        "operationId": "get_diagnosis_run_api_v1_diagnosis_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis-runs/{run_id}/report": {
      "get": {
        "tags": [
          "diagnosis",
          "diagnosis"
        ],
        "summary": "Get Diagnosis Report",
        "operationId": "get_diagnosis_report_api_v1_diagnosis_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis-runs/{run_id}/trace": {
      "get": {
        "tags": [
          "diagnosis",
          "diagnosis"
        ],
        "summary": "Get Diagnosis Trace",
        "operationId": "get_diagnosis_trace_api_v1_diagnosis_runs__run_id__trace_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisTraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis-runs/{run_id}/events": {
      "get": {
        "tags": [
          "diagnosis",
          "diagnosis"
        ],
        "summary": "Stream Diagnosis Events",
        "description": "Replay persisted run events and follow them briefly over SSE.\n\nThe database event log is the source of truth. A reconnect starts after\n``Last-Event-ID`` and therefore cannot lose events already committed by a\nworker. The stream ends after a terminal event or a bounded idle window;\nclients can reconnect with the last sequence number.",
        "operationId": "stream_diagnosis_events_api_v1_diagnosis_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{evidence_code}": {
      "get": {
        "tags": [
          "evidence",
          "evidence"
        ],
        "summary": "Get Evidence",
        "operationId": "get_evidence_api_v1_evidence__evidence_code__get",
        "parameters": [
          {
            "name": "evidence_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}/download-url": {
      "get": {
        "tags": [
          "artifacts",
          "artifacts"
        ],
        "summary": "Get Artifact Download Url",
        "operationId": "get_artifact_download_url_api_v1_artifacts__artifact_id__download_url_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDownloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artifacts/{artifact_id}/content": {
      "get": {
        "tags": [
          "artifacts",
          "artifacts"
        ],
        "summary": "Get Artifact Content",
        "operationId": "get_artifact_content_api_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactDownloadResponse": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artifact Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_seconds": {
            "type": "integer",
            "title": "Expires Seconds"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "url",
          "expires_seconds",
          "content_type"
        ],
        "title": "ArtifactDownloadResponse"
      },
...
      "DiagnosisRunEventSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "payload": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sequence",
          "event_type",
          "stage",
          "message",
          "payload",
          "created_at"
        ],
        "title": "DiagnosisRunEventSchema"
      },
      "DiagnosisRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Incident Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "ontology_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ontology Version"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "celery_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celery Task Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "estimated_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "incident_id",
          "idempotency_key",
          "status",
          "model_provider",
          "model_name",
          "prompt_version",
          "ontology_version",
          "attempt_number",
          "celery_task_id",
          "started_at",
          "finished_at",
          "error_type",
          "error_message",
          "total_duration_ms",
          "input_tokens",
          "output_tokens",
          "estimated_cost",
          "created_at",
          "updated_at"
        ],
        "title": "DiagnosisRunResponse",
        "description": "GET /diagnosis-runs/{id} response \u2014 run status only, no report payload."
      },
...
      "DiagnosisTraceResponse": {
        "properties": {
          "diagnosis_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnosis Run Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/DiagnosisRunEventSchema"
            },
            "type": "array",
            "title": "Events"
          },
          "tool_executions": {
            "items": {
              "$ref": "#/components/schemas/ToolExecutionSchema"
            },
            "type": "array",
            "title": "Tool Executions"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "diagnosis_run_id",
          "events",
          "tool_executions",
          "evidence"
        ],
        "title": "DiagnosisTraceResponse"
      },
...
      "EvidenceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "evidence_code": {
            "type": "string",
            "title": "Evidence Code"
          },
          "tool_execution_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tool Execution Id"
          },
          "evidence_type": {
            "type": "string",
            "title": "Evidence Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "metrics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "filters": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "artifact_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "evidence_code",
          "tool_execution_id",
          "evidence_type",
          "title",
          "summary",
          "metrics",
          "filters",
          "artifact_id",
          "created_at"
        ],
        "title": "EvidenceResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "diagnosis_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnosis Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "total_estimated_lost_intents": {
            "type": "integer",
            "title": "Total Estimated Lost Intents"
          },
          "explained_lost_intents": {
            "type": "integer",
            "title": "Explained Lost Intents"
          },
          "unexplained_lost_intents": {
            "type": "integer",
            "title": "Unexplained Lost Intents"
          },
          "missing_data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Data"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "root_causes": {
            "items": {
              "$ref": "#/components/schemas/RootCauseSchema"
            },
            "type": "array",
            "title": "Root Causes"
          },
          "ontology_version": {
            "type": "string",
            "title": "Ontology Version"
          },
          "validator_version": {
            "type": "string",
            "title": "Validator Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "alternative_explanations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternative Explanations"
          }
        },
        "type": "object",
        "required": [
          "id",
          "diagnosis_run_id",
          "status",
          "summary",
          "total_estimated_lost_intents",
          "explained_lost_intents",
          "unexplained_lost_intents",
          "missing_data",
          "recommended_actions",
          "root_causes",
          "ontology_version",
          "validator_version",
          "created_at"
        ],
        "title": "ReportResponse",
        "description": "GET /diagnosis-runs/{id}/report response."
      },
      "RootCauseSchema": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Confidence"
          },
          "estimated_lost_intents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Lost Intents"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "evidence_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Codes"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "label",
          "category",
          "confidence",
          "estimated_lost_intents",
          "explanation",
          "evidence_codes",
          "rank"
        ],
        "title": "RootCauseSchema"
      },
...
      "ToolExecutionSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tool_call_id": {
            "type": "string",
            "title": "Tool Call Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "artifact_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tool_call_id",
          "tool_name",
          "status",
          "duration_ms",
          "row_count",
          "artifact_id",
          "error_type",
          "error_message"
        ],
        "title": "ToolExecutionSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}